Calculate distance from prisons to all airports within x distance
This script calculates both the objects (airports) within 5000 meters of each prison and their distance in meters to the prison. It creates a new column in the prisons data for each object and each object distance. The airport objects will be ordered into the columns based on their distance to the prison, so column 1 will be the closest, column 2 will be the second closest, and so on.